tieded code up with UIBarButtonItem extension
