+++ WBooks/Library/MainMenuController.swift
@@ -40,15 +40,9 @@ class MainMenuController: UIViewController {
         self.title = "LIBRARY"
         nav?.titleTextAttributes = [NSAttributedStringKey.foregroundColor: UIColor.white]
         //Left notification button
-        let notificationsButton = UIButton(type: .system)
-        notificationsButton.setImage(#imageLiteral(resourceName: "ic_notifications").withRenderingMode(.alwaysOriginal), for: .normal)
-        notificationsButton.frame = CGRect(x: 0, y: 0, width: 34, height: 34)
-        navigationItem.leftBarButtonItem = UIBarButtonItem(customView: notificationsButton)
+        navigationItem.leftBarButtonItem = UIBarButtonItem.notificationButton()
         //Right search button
-        let searchButton = UIButton(type: .system)
-        searchButton.setImage(#imageLiteral(resourceName: "ic_search.png").withRenderingMode(.alwaysOriginal), for: .normal)
-        searchButton.frame = CGRect(x: 0, y: 0, width: 34, height: 34)
-        navigationItem.rightBarButtonItem = UIBarButtonItem(customView: searchButton)
+        navigationItem.rightBarButtonItem = UIBarButtonItem.searchButton()
     }
     
     private func configureTableView() {
@@ -90,3 +84,4 @@ extension MainMenuController: UITableViewDataSource, UITableViewDelegate {
         return 100
     }
 }
+
